refactor: Remove dead engine, rename ParallelEngine to TracerouteEngine
The old sequential TracerouteEngine (engine.rs) was unused since
api.rs only calls ParallelEngine. Delete it and promote
ParallelEngine to be the sole engine:

- Delete old engine.rs (dead code since v0.4.0)
- Rename parallel_engine.rs -> engine.rs
- Rename ParallelEngine -> TracerouteEngine
- Update CHANGELOG with v0.7.0 changes
