+++ src/traceroute.rs
@@ -27,7 +27,6 @@ pub mod config;
 pub mod engine;
 pub mod error;
 pub mod isp_from_path;
-pub mod parallel_engine;
 pub mod result;
 pub mod types;
